Added methods in utils
